refactor: decouple city alerts and news API logic into dedicated service modules
