+++ client/src/app/globals.css
@@ -53,7 +53,6 @@ textarea {
   font: inherit;
 }
 
-
 input,
 select,
 textarea {
