+++ client/src/components/ui/button.tsx
@@ -1,4 +1,3 @@
-
 import * as React from 'react';
 import { Slot } from '@radix-ui/react-slot';
 import { cn } from '@/lib/utils';
