+++ client/src/components/ui/label.tsx
@@ -1,4 +1,3 @@
-
 import * as React from 'react';
 import * as LabelPrimitive from '@radix-ui/react-label';
 import { cn } from '@/lib/utils';
